Timeline + News + Routes + Publications + Comments + Discussions + Messages

+++ app/controllers/User2DiscussionREST.java
@@ -6,11 +6,9 @@ import java.util.List;
 
 import models.Discussion;
 import models.Message;
-import models.Publication;
 import models.User;
 import models.User2Discussion;
 import models.ApiAppUser;
-import models.Wall;
 
 import controllers.ApiUserREST;
 
@@ -23,17 +21,53 @@ import play.mvc.Result;
 
 public class User2DiscussionREST extends ItemREST {
 
-	public static Result getDiscussions() {
-
+	public static Result all() {
 		final User user = Application.getLocalUser(session());
 		if (user == null){
 			return badRequest(Constants.USER_NOT_LOGGED_IN.toString());
 		}
-		User2Discussion user2disc = User2Discussion.findById(user.id);
+		User2Discussion user2disc = User2Discussion.findById(user.id, User2Discussion.class);
 		if (user2disc == null) {
 			return badRequest(Constants.DISCUSSIONS_EMPTY.toString());
 		}
 		List<Discussion> discussions = user2disc.all();
+		if (discussions.size() == 0) {
+			return badRequest(Constants.DISCUSSIONS_EMPTY.toString());
+		} else {
+			// Return the response
+			return ok(Json.toJson(Discussion.discussionsToObjectNodes(discussions)));
+		}
+	}
+	
+	public static Result getDiscussions() {
+		final User user = Application.getLocalUser(session());
+		if (user == null){
+			return badRequest(Constants.USER_NOT_LOGGED_IN.toString());
+		}
+		User2Discussion user2disc = User2Discussion.findById(user.id, User2Discussion.class);
+		if (user2disc == null) {
+			return badRequest(Constants.DISCUSSIONS_EMPTY.toString());
+		}
+		List<Discussion> discussions = user2disc.getDiscussions(0, 10);
+		if (discussions.size() == 0) {
+			return badRequest(Constants.DISCUSSIONS_EMPTY.toString());
+		} else {
+			// Return the response
+			return ok(Json.toJson(Discussion.discussionsToObjectNodes(discussions)));
+		}
+	}
+	
+	
+	public static Result getRangeDiscussions(int from, int to) {
+		final User user = Application.getLocalUser(session());
+		if (user == null){
+			return badRequest(Constants.USER_NOT_LOGGED_IN.toString());
+		}
+		User2Discussion user2disc = User2Discussion.findById(user.id, User2Discussion.class);
+		if (user2disc == null) {
+			return badRequest(Constants.DISCUSSIONS_EMPTY.toString());
+		}
+		List<Discussion> discussions = user2disc.getDiscussions(from, to);
 		if (discussions.size() == 0) {
 			return badRequest(Constants.DISCUSSIONS_EMPTY.toString());
 		} else {
@@ -48,7 +82,7 @@ public class User2DiscussionREST extends ItemREST {
 		if (user == null){
 			return badRequest(Constants.USER_NOT_LOGGED_IN.toString());
 		}
-		User2Discussion user2disc = User2Discussion.findById(user.id);
+		User2Discussion user2disc = User2Discussion.findById(user.id, User2Discussion.class);
 		if (user2disc == null) {
 			return badRequest(Constants.DISCUSSIONS_EMPTY.toString());
 		}
@@ -61,13 +95,13 @@ public class User2DiscussionREST extends ItemREST {
 		}
 	}
 
-	
+
 	public static Result getDiscussion(String discussionId) {
 		final User user = Application.getLocalUser(session());
 		if (user == null){
 			return badRequest(Constants.USER_NOT_LOGGED_IN.toString());
 		}
-		User2Discussion user2disc = User2Discussion.findById(user.id);
+		User2Discussion user2disc = User2Discussion.findById(user.id, User2Discussion.class);
 		if (user2disc == null) {
 			return badRequest(Constants.DISCUSSIONS_EMPTY.toString());
 		}
@@ -79,7 +113,7 @@ public class User2DiscussionREST extends ItemREST {
 		return ok(Json.toJson(Discussion.discussionToFullObjectNode(discussion)));
 	}
 
-	
+
 	public static Result getMessage(String messageId) {
 		final User user = Application.getLocalUser(session());
 		if (user == null){
@@ -89,7 +123,7 @@ public class User2DiscussionREST extends ItemREST {
 		if(json == null) {
 			return badRequest(Constants.JSON_EMPTY.toString());
 		}
-		User2Discussion user2disc = User2Discussion.findById(user.id);
+		User2Discussion user2disc = User2Discussion.findById(user.id, User2Discussion.class);
 		if (user2disc == null) {
 			return badRequest(Constants.DISCUSSIONS_EMPTY.toString());
 		}
@@ -107,6 +141,7 @@ public class User2DiscussionREST extends ItemREST {
 		return ok(Json.toJson(Message.messageToObjectNode(message)));
 	}
 
+	
 	public static Result updateMessage(String messageId) {
 		final User user = Application.getLocalUser(session());
 		if (user == null){
@@ -116,7 +151,7 @@ public class User2DiscussionREST extends ItemREST {
 		if(json == null) {
 			return badRequest(Constants.JSON_EMPTY.toString());
 		}
-		User2Discussion user2disc = User2Discussion.findById(user.id);
+		User2Discussion user2disc = User2Discussion.findById(user.id, User2Discussion.class);
 		if (user2disc == null) {
 			return badRequest(Constants.DISCUSSIONS_EMPTY.toString());
 		}
@@ -137,7 +172,8 @@ public class User2DiscussionREST extends ItemREST {
 		List <ObjectId> otherUsers = discussion.userIds;
 		otherUsers.remove(user.id);
 
-		setUsersDiscussionUnread(discussion, otherUsers);
+		// Set the other users discusion to UNREAD
+		setUsersDiscussionToUnread(discussion, otherUsers);
 
 		// Return a copy of the message
 		return ok(Json.toJson(Message.messageToObjectNode(message)));
@@ -160,11 +196,10 @@ public class User2DiscussionREST extends ItemREST {
 		List<User2Discussion> user2discList = new ArrayList<User2Discussion>(); 
 
 		discussion = new Discussion();		// Create discussion
-		discussion.messageIds = new ArrayList<ObjectId>();
 		discussion.userIds = new ArrayList<ObjectId>();
 		discussion.subject = json.findPath("subject").getTextValue();
 		discussion.save();
-		
+
 		try {
 
 			message = new Message();		// Create message
@@ -172,18 +207,19 @@ public class User2DiscussionREST extends ItemREST {
 			message.writerId = user.id;
 			message.save();
 
-			discussion.addMessage(message); // Add message to discussion
+			discussion.firstMessage = message.id; // This is the discussions first message
+			discussion.save();
 
 			// Add discussion to all readers
 			Iterator<JsonNode> userIds = json.findPath("users").getElements();
-			
+
 			while(userIds.hasNext()){
 				String userId = userIds.next().toString().replace("\"", "");
 
 				User receiver = User.findById(userId, User.class);
 
 				if (receiver != null){
-					User2Discussion user2disc = User2Discussion.findById(receiver.id);
+					User2Discussion user2disc = User2Discussion.findById(receiver.id, User2Discussion.class);
 
 					// If is users first discussion, create new User2Discussion
 					if (user2disc == null){
@@ -193,24 +229,24 @@ public class User2DiscussionREST extends ItemREST {
 						/****************************************/
 						user2disc.save();
 					}
-					
+
 					if (user2disc.discussionIds == null){
 						user2disc.discussionIds = new ArrayList<ObjectId>();
 					}
-					
+
 					if (user2disc.unread == null){
 						user2disc.unread = new ArrayList<ObjectId>();
 					}
 
-					discussion.addUser(receiver);
-					user2disc.addDiscussion(discussion); // Add discussions id to this user
+					discussion.addUser(receiver); //  Add user to discussion
+					user2disc.addDiscussion(discussion); // Add discussion to this user
 					user2disc.setRead(discussion, false);
 					user2discList.add(user2disc);
 				}
 			}
 
 			// Add discussion to sender
-			User2Discussion user2disc = User2Discussion.findById(user.id);
+			User2Discussion user2disc = User2Discussion.findById(user.id, User2Discussion.class);
 
 			// If is users first discussion, create new User2Discussion
 			if (user2disc == null){
@@ -220,24 +256,24 @@ public class User2DiscussionREST extends ItemREST {
 				/****************************************/
 				user2disc.save();
 			}
-			
+
 			if (user2disc.discussionIds == null){
 				user2disc.discussionIds = new ArrayList<ObjectId>();
 			}
-			
+
 			if (user2disc.unread == null){
 				user2disc.unread = new ArrayList<ObjectId>();
 			}
 
-			discussion.addUser(user);
-			user2disc.addDiscussion(discussion);
+			discussion.addUser(user); // Add user to discussion
+			user2disc.addDiscussion(discussion); // Add discussion to user
 			user2disc.setRead(discussion, true);
 			user2discList.add(user2disc);
 
 		} catch (Exception e) {
 
 			discussion.delete();
-			
+
 			// TODO REMOVE DISCUSSION, MESSAGE AND USER2DISCUSSIONS
 
 			return badRequest(Constants.JSON_EMPTY.toString());
@@ -252,7 +288,7 @@ public class User2DiscussionREST extends ItemREST {
 		return ok(Json.toJson(Discussion.discussionToFullObjectNode(discussion)));
 	}
 
-	
+
 	public static Result deleteDiscussion(String id){
 		final User user = Application.getLocalUser(session());
 		if (user == null){
@@ -262,7 +298,7 @@ public class User2DiscussionREST extends ItemREST {
 		if(json == null) {
 			return badRequest(Constants.JSON_EMPTY.toString());
 		}
-		User2Discussion user2disc = User2Discussion.findById(user.id);
+		User2Discussion user2disc = User2Discussion.findById(user.id, User2Discussion.class);
 		if (user2disc == null || !user2disc.discussionIds.contains(new ObjectId(id))){
 			return badRequest(Constants.DISCUSSIONS_EMPTY.toString());
 		}
@@ -272,12 +308,14 @@ public class User2DiscussionREST extends ItemREST {
 		}
 
 		user2disc.removeDiscussion(discussion);
+		user2disc.save();
+
 		discussion.removeUser(user);
 		discussion.save();
 
 		if (discussion.userIds.isEmpty()){
 
-			for(ObjectId oid : discussion.messageIds){
+			for(ObjectId oid : discussion.repliesIds){
 				Message message = discussion.findMessageById(oid);
 				message.delete();
 			}
@@ -286,7 +324,7 @@ public class User2DiscussionREST extends ItemREST {
 		}
 		return ok(Json.toJson(Discussion.discussionToFullObjectNode(discussion)));
 	}
-	
+
 
 	public static Result reply() {
 		final User user = Application.getLocalUser(session());
@@ -297,7 +335,7 @@ public class User2DiscussionREST extends ItemREST {
 		if(json == null) {
 			return badRequest(Constants.JSON_EMPTY.toString());
 		}
-		User2Discussion user2disc = User2Discussion.findById(user.id);
+		User2Discussion user2disc = User2Discussion.findById(user.id, User2Discussion.class);
 		if (user2disc == null) {
 			return badRequest(Constants.DISCUSSIONS_EMPTY.toString());
 		}
@@ -319,13 +357,13 @@ public class User2DiscussionREST extends ItemREST {
 		List <ObjectId> otherUsers = discussion.userIds;
 		otherUsers.remove(user.id);
 
-		setUsersDiscussionUnread(discussion, otherUsers);
+		setUsersDiscussionToUnread(discussion, otherUsers);
 
 		// Return a copy of the discussion
 		return ok(Json.toJson(Message.messageToObjectNode(message)));
 	}
 
-	
+
 	public static Result replyToMessage(String id){
 		final User user = Application.getLocalUser(session());
 		if (user == null){
@@ -335,7 +373,7 @@ public class User2DiscussionREST extends ItemREST {
 		if(json == null) {
 			return badRequest(Constants.JSON_EMPTY.toString());
 		}
-		User2Discussion user2disc = User2Discussion.findById(user.id);
+		User2Discussion user2disc = User2Discussion.findById(user.id, User2Discussion.class);
 		if (user2disc == null) {
 			return badRequest(Constants.DISCUSSIONS_EMPTY.toString());
 		}
@@ -361,15 +399,15 @@ public class User2DiscussionREST extends ItemREST {
 			List <ObjectId> otherUsers = discussion.userIds;
 			otherUsers.remove(user.id);
 
-			setUsersDiscussionUnread(discussion, otherUsers);
+			setUsersDiscussionToUnread(discussion, otherUsers);
 
 			// Return a copy of the discussion
 			return ok(Json.toJson(Message.messageToObjectNode(message)));
 		}
 		return badRequest(Constants.MESSAGES_EMPTY.toString());
 	}
-	
-	
+
+
 	public static Result deleteMessage(String id){
 		final User user = Application.getLocalUser(session());
 		if (user == null){
@@ -379,7 +417,7 @@ public class User2DiscussionREST extends ItemREST {
 		if(json == null) {
 			return badRequest(Constants.JSON_EMPTY.toString());
 		}
-		User2Discussion user2disc = User2Discussion.findById(user.id);
+		User2Discussion user2disc = User2Discussion.findById(user.id, User2Discussion.class);
 		if (user2disc == null) {
 			return badRequest(Constants.DISCUSSIONS_EMPTY.toString());
 		}
@@ -392,22 +430,29 @@ public class User2DiscussionREST extends ItemREST {
 		if (message == null){
 			return badRequest(Constants.MESSAGES_EMPTY.toString());
 		}
+		
+		// If someone is trying to delete the first message from a discussion
+		// That cant be done
+		if (message.id.equals(discussion.firstMessage)){
+			return forbidden(Constants.UNAUTHORIZED.toString());
+		}
+		
 		discussion.deleteMessage(message);
 		discussion.save();
 		message.delete();
-		
+
 		return ok(Json.toJson(Message.messageToObjectNode(message)));
 	}
 
-	
+
 	/** Change discussion to unread for this users
 	 * @param discussionId
 	 * @param userIds
 	 */
-	public static void setUsersDiscussionUnread(Discussion discussion, List<ObjectId> userIds){
+	public static void setUsersDiscussionToUnread(Discussion discussion, List<ObjectId> userIds){
 
 		for (ObjectId oid : userIds){
-			User2Discussion user2disc = User2Discussion.findById(oid);
+			User2Discussion user2disc = User2Discussion.findById(oid, User2Discussion.class);
 
 			// If is users first discussion, create new User2Discussion
 			if (user2disc == null){
@@ -428,7 +473,7 @@ public class User2DiscussionREST extends ItemREST {
 			return badRequest(Constants.USER_NOT_LOGGED_IN.toString());
 		}
 
-		User2Discussion user2disc = User2Discussion.findById(user.id);
+		User2Discussion user2disc = User2Discussion.findById(user.id, User2Discussion.class);
 		if (user2disc == null) {
 			return badRequest(Constants.DISCUSSIONS_EMPTY.toString());
 		}
