+++ app/models/Discussion.java
@@ -26,12 +26,19 @@ public class Discussion extends Post {
 	@Override
 	public void delete() {
 		for (ObjectId oid : userIds){
-			User2Discussion user2disc = User2Discussion.findById(oid);
+			User2Discussion user2disc = User2Discussion.findById(oid, User2Discussion.class);
 			user2disc.removeDiscussion(this);
 		}
 
-		for(ObjectId oid : messageIds){
-			Message message = Message.findById(oid, Message.class);
+		if (repliesIds != null){
+			for(ObjectId oid : repliesIds){
+				Message message = Message.findById(oid, Message.class);
+				message.delete();
+			}
+		}
+
+		Message message = Message.findById(firstMessage, Message.class);
+		if (message != null){
 			message.delete();
 		}
 
@@ -55,6 +62,7 @@ public class Discussion extends Post {
 			userIds.remove(user.id);
 			if (userIds.isEmpty()){
 				userIds = null;
+				this.delete();
 			}
 		}
 		this.save();
@@ -81,9 +89,17 @@ public class Discussion extends Post {
 		discussionNode.put("id", discussion.id.toString());
 		discussionNode.put("unread", 0);
 		discussionNode.put("subject", discussion.subject);
-		discussionNode.put("users", Json.toJson(User.usersSmallInfo(discussion.userIds)));
+		discussionNode.put("users", Json.toJson(User.userIdsToShortObjectNode(discussion.userIds)));
 		discussionNode.put("timeStamp", discussion.id.getTime());
 		discussionNode.put("lastWrote", discussion.lastWrote.toString());
+
+		Message message = Message.findById(discussion.firstMessage, Message.class);
+		if (message != null){
+			discussionNode.put("content", message.message);
+		} else {
+			discussion.delete();
+		}
+
 		return discussionNode;
 	}
 
@@ -96,10 +112,18 @@ public class Discussion extends Post {
 		discussionNode.put("id", discussion.id.toString());
 		discussionNode.put("unread", 0);
 		discussionNode.put("subject", discussion.subject);
-		discussionNode.put("users", Json.toJson(User.usersSmallInfo(discussion.userIds)));
+		discussionNode.put("users", Json.toJson(User.userIdsToShortObjectNode(discussion.userIds)));
 		discussionNode.put("timeStamp", discussion.id.getTime());
 		discussionNode.put("lastWrote", discussion.lastWrote.toString());
 		discussionNode.put("messages", Json.toJson(Message.messagesToObjectNodes(discussion.getMessages())));
+
+		Message message = Message.findById(discussion.firstMessage, Message.class);
+		if (message != null){
+			discussionNode.put("content", message.message);
+		} else {
+			discussion.delete();
+		}
+
 		return discussionNode;
 	}
 }
